+++ docsig/messages.py
@@ -8,13 +8,13 @@ W101 = "{module}::{func} is missing a docstring"
 
 #: Exxx: Error
 #: E1xx: Docstring
-E101 = "parameters out of order"
-E102 = "includes parameters that do not exist"
-E103 = "parameters missing"
-E104 = "return statement documented for None"
-E105 = "return missing from docstring"
-E106 = "duplicate parameters found"
-E107 = "parameter appears to be incorrectly documented"
+E101 = "E101: parameters out of order"
+E102 = "E102: includes parameters that do not exist"
+E103 = "E103: parameters missing"
+E104 = "E104: return statement documented for None"
+E105 = "E105: return missing from docstring"
+E106 = "E106: duplicate parameters found"
+E107 = "E107: parameter appears to be incorrectly documented"
 
 #: Hxxx: Hint
 #: H1xx: Docstring