+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = "2022, Stephen Whitlock"
 author = "Stephen Whitlock"
 
 # The full version, including alpha/beta/rc tags
-release = "0.12.0"
+release = "0.12.1"
 
 
 # -- General configuration ---------------------------------------------------