+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Adds codes to printed messages
 
+### Changed
+- Fails checks with bad indentation
+
 [0.9.0](https://github.com/jshwi/docsig/releases/tag/v0.9.0) - 2022-06-26
 ------------------------------------------------------------------------
 ### Added