+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.11.0...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- Allows for params with any amount of spaces before description
 
 [0.11.0](https://github.com/jshwi/docsig/releases/tag/v0.11.0) - 2022-06-27
 ------------------------------------------------------------------------