+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/docsig/compare/v0.12.0...HEAD)
+[Unreleased](https://github.com/jshwi/docsig/compare/v0.12.1...HEAD)
+------------------------------------------------------------------------
+
+[0.12.1](https://github.com/jshwi/docsig/releases/tag/v0.12.1) - 2022-06-29
 ------------------------------------------------------------------------
 ### Fixed
 - Prevents duplicate error messages from occurring